feat: change signUpFlow and add find Account, get new password

+++ rentalService/src/main/java/com/itemrental/rentalService/service/AuthService.java
@@ -1,7 +1,9 @@
 package com.itemrental.rentalService.service;
 
+import com.itemrental.rentalService.entity.ResetToken;
 import com.itemrental.rentalService.entity.User;
 import com.itemrental.rentalService.entity.VerificationToken;
+import com.itemrental.rentalService.repository.ResetTokenRepository;
 import com.itemrental.rentalService.repository.UserRepository;
 import com.itemrental.rentalService.repository.VerificationTokenRepository;
 import lombok.RequiredArgsConstructor;
@@ -15,15 +17,30 @@ import java.util.Optional;
 @Slf4j
 public class AuthService {
     private final VerificationTokenRepository verificationTokenRepository;
+    private final ResetTokenRepository resetTokenRepository;
     private final UserRepository userRepository;
 
-    public void verifyToken(String token) {
+    public boolean verifyVerificationToken(String token) {
         Optional<VerificationToken> opVerificationToken = verificationTokenRepository.findById(token);
-        VerificationToken verificationToken = opVerificationToken.orElseThrow(() -> new IllegalArgumentException(""));
+        if(opVerificationToken.isEmpty()) return false;
+        VerificationToken verificationToken = opVerificationToken.orElseThrow(() -> new IllegalArgumentException("잘못된 인증 토큰"));
         Optional<User> opUser = userRepository.findByEmail(verificationToken.getEmail());
+        if(opUser.isEmpty()) return false;
         User user = opUser.orElseThrow(() -> new IllegalArgumentException("없는 사용자"));
-        user.setEmailVerified(true);
+        user.setUserState(User.UserState.PENDING_PROFILE_SETUP);
         verificationTokenRepository.deleteById(token);
         userRepository.save(user);
+        return true;
+    }
+
+    public boolean verifyResetToken(String token){
+        Optional<ResetToken> opResetToken = resetTokenRepository.findById(token);
+        if(opResetToken.isEmpty()) return false;
+        ResetToken resetToken = opResetToken.orElseThrow(() -> new IllegalArgumentException("잘못된 비밀번호 초기화 토큰"));
+        Optional<User> opUser = userRepository.findByEmail(resetToken.getEmail());
+        if(opUser.isEmpty()){
+            return false;
+        }
+        return true;
     }
 }
